[Gameplay] Pellets are now attracted to pac-man.
* When near, Pac-Man will attract pellets.

Note: The raycasts are done in update. After much experimenting (trying
to get smooth and correct motions when doing raycasts in FixedUpdate +
moving in update was not trivial) and checking with different online
posts + doing some trivial performance checks it does not seem to be a
burden. So, unless there is noticable performance issues when running
this on a another computer I think I will stick to this solution.