ShockWave shader added
Now we can create a shockwave animation with a shader. This shader will
replace the self illuminated shader later for the floor because I have
included self illumination and texture mapping into this shader. The
animation is controlled through a script called ShockWave that will
update the radius of the shock wave and trigger it on a key press. Right
now that key is F4. This should later be used in conjuction with the EMP
effect.